Daily work

+++ database/migrations/2021_04_12_090135_add_roles_users_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Helpers\Roles;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -13,17 +14,10 @@ class AddRolesUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->tinyInteger('roles')->after('email')->default(\App\Helpers\Roles::GUEST);
-        });
-
-        try {
-            \App\Models\User::whereEmail('support@cryptomatix.io')->update(
-                [
-                    'roles' => \App\Helpers\Roles::SU
-                ]
-            );
-        } catch (Exception $exception) {
+        if (!Schema::hasColumn('users', 'roles')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->tinyInteger('roles')->after('email')->default(Roles::GUEST);
+            });
         }
     }
 
@@ -34,8 +28,10 @@ class AddRolesUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('roles');
-        });
+        if (Schema::hasColumn('users', 'roles')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('roles');
+            });
+        }
     }
 }
