+++ database/migrations/2021_04_13_125621_add_otp_token_to_users_table.php
@@ -14,12 +14,14 @@ class AddOtpTokenToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('otp_token')->nullable()->after('remember_token');
-            $table->enum('token_status', 
-                [User::INACTIVED_TOKEN, User::ACTIVED_TOKEN]
-            )->default(User::INACTIVED_TOKEN)->after('otp_token');
-        });
+        if (!Schema::hasColumns('users', ['otp_token', 'token_status'])) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('otp_token')->nullable()->after('remember_token');
+                $table->enum('token_status',
+                    [User::INACTIVED_TOKEN, User::ACTIVED_TOKEN]
+                )->default(User::INACTIVED_TOKEN)->after('otp_token');
+            });
+        }
     }
 
     /**
@@ -29,9 +31,11 @@ class AddOtpTokenToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('otp_token');
-            $table->dropColumn('token_status');
-        });
+        if (Schema::hasColumns('users', ['otp_token', 'token_status'])) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('otp_token');
+                $table->dropColumn('token_status');
+            });
+        }
     }
 }
