+++ database/migrations/2021_04_13_141558_add_first_login_to_users_table.php
@@ -14,11 +14,13 @@ class AddFirstLoginToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->enum('first_login', 
-                [User::NO, User::YES]
-            )->default(User::YES)->after('token_status');
-        });
+        if (!Schema::hasColumn('users', 'first_login')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->enum('first_login',
+                    [User::NO, User::YES]
+                )->default(User::YES)->after('token_status');
+            });
+        }
     }
 
     /**
@@ -28,8 +30,10 @@ class AddFirstLoginToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('first_login');
-        });
+        if (Schema::hasColumn('users', 'first_login')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('first_login');
+            });
+        }
     }
 }
